refactor: centralize comment utilities
